requirements.txt
playwright==1.44.0
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
python-telegram-bot==21.3
aiosqlite==0.20.0
SQLAlchemy==2.0.30
rapidfuzz==3.9.3
loguru==0.7.2
selectolax>=0.3.21
